implement RTK for general state
