--- mobile/src/redux/reducers/faq-screen/faq-data.ts
@@ -1,105 +1,0 @@
-export type FAQData = Array<{
-  title: string;
-  questions: Array<{
-    question: string;
-    answer: string;
-  }>;
-}>;
-
-export const faqData: FAQData = [
-  {
-    title: "General",
-    questions: [
-      {
-        question: "What exactly is DzCode i/o ?",
-        answer: `An open-source online community of algerian developers, no organization or any political party is involved, only us dz developers.`,
-      },
-      {
-        question: "What are the Goals of DzCode i/o ?",
-        answer: `
-- Share our experiences (in form of [Articles](/Articles)), there are many algerian developers that accumulated lots experiences, dzcode.io gives them the chance to share it with the right audience.
-- Guide new developers to build their career (with detailed [documentation](/Learn)), and understand the software market, instead of wasting their time on other useless resources.
-- Fix some common algerian software problems, in form of ready to use open-source [softwares](/Projects) (packages, libraries ...etc).`,
-      },
-      {
-        question: "What benefit will i get from it ?",
-        answer: `There are couple of benefits, to name few:
-
-- You will meet other algerian developers, experienced and juniors, where you can grow your network!
-- You'll get to experience a professional environment, which is completely different than the one you used to in school.
-- Your contribution will be shown on your Github profile, and that is a big plus in your professional career!
-`,
-      },
-      {
-        question: "dzcode, dzcode.io or DzCode i/o ?",
-        answer: `You may get confused with the different terms, so here's the official breakdown:
-
-- **DzCode i/o** (with capital C) is the official name of the community.
-- **dzcode.io** is the website of the DzCode i/o, this has been voted for [here](https://www.facebook.com/groups/dzdevs/permalink/3052880478141052/).
-- **dzcode** is a short name, not official, but it's usually used to refer DzCode i/o.
-`,
-      },
-    ],
-  },
-  {
-    title: "Participation",
-    questions: [
-      {
-        question: "How to join DzCode i/o ?",
-        answer: `We're most active in slack, it's highly recommended to [join us there](https://join.slack.com/t/dzcode/shared_invite/zt-ek9kscb7-m8z_~cBjX79l~uchuABPFQ).`,
-      },
-      {
-        question: "I want to code, where should I start ?",
-        answer: `Go to [/Contribute](/Contribute) page, you will see list projects with their open tasks, pick whatever you like and start coding 🔥.
-
-**Note** : Please make sure to inform the project maintainer to reserve the task for you, you can do that by commenting on the issue ticket.`,
-      },
-      {
-        question: "I like the idea, what are the different ways i can contribute to DzCode i/o ?",
-        answer: `Apart from checking out the [/Contribute](/Contribute) page, you can also contribute to DzCode i/o by:
-
-- Adding your open-source projects to dzcode.io website, you can do that [here](/Learn/About_dzcode_io/Add_Your_Project_To_dzcode_io).
-- Writing about your experience in form of articles, it's as easy as following the steps [here](/Learn/About_dzcode_io/Add_Your_Article_To_dzcode_io).
-`,
-      },
-    ],
-  },
-  {
-    title: "Articles",
-    questions: [
-      {
-        question: "How to write an article in dzcode.io",
-        answer: `See this [tutorial](/Learn/About_dzcode_io/Add_Your_Article_To_dzcode_io).`,
-      },
-      {
-        question: "The Articles should they be in English ?",
-        answer:
-          "Yes, to reach a wide range of audience we will add a feature to allow article localization and multiple language articles in the future.",
-      },
-      {
-        question: "Can I write about anything ?",
-        answer: "As long as its IT related yes you can write about anything.",
-      },
-    ],
-  },
-  {
-    title: "Projects",
-    questions: [
-      {
-        question: "How to add my open-source project to dzcode.io",
-        answer: `See this [tutorial](/Learn/About_dzcode_io/Add_Your_Project_To_dzcode_io).`,
-      },
-      {
-        question: "What makes my project eligible to be added to dzcode.io",
-        answer: "It has to solve an Algerian problem, or, to be written by an Algerian Developer.",
-      },
-      {
-        question: "Why should i add my project to dzcode.io ?",
-        answer: `
-- Your project will potentially get noticed by more algerian developers.
-- You will get potential contribution from other DZ developers, and eventually better shape your piece of software.
-`,
-      },
-    ],
-  },
-];
