+++ mobile/src/screens/projects/card-item/index.tsx
@@ -3,8 +3,7 @@ import React, { FC, memo } from "react";
 import { Button, Card, Paragraph, Title } from "react-native-paper";
 import { useSelector } from "react-redux";
 
-import { StateInterface } from "../../../redux";
-import { GeneralState } from "../../../redux/reducers/general";
+import { selectTheme } from "../../../store/general/selectors/theme";
 import { Colors } from "../../../styles/colors";
 import { openLink } from "../../../utils/link";
 import { cardStyles } from "./styles";
@@ -16,7 +15,7 @@ interface CardItemProps {
 const CardItem: FC<CardItemProps> = ({
   project: { title, description, githubURI, image },
 }: CardItemProps) => {
-  const { theme } = useSelector<StateInterface, GeneralState>((state) => state.general);
+  const theme = useSelector(selectTheme);
   return (
     <Card style={cardStyles.mainView}>
       <Card.Cover
