+++ mobile/src/store/general/selectors/index.ts
@@ -1,0 +1,3 @@
+import { RootState } from "../..";
+
+export const domain = (state: RootState) => state.general;
