+++ mobile/src/store/general/selectors/theme.ts
@@ -1,0 +1,5 @@
+import { createSelector } from "@reduxjs/toolkit";
+
+import { domain } from ".";
+
+export const selectTheme = createSelector(domain, (state) => state.theme);
